Add several folders and modules #1
