+++ src/Config.hs
@@ -3,29 +3,19 @@
 
 module Config where
 
-import Control.Concurrent (threadDelay)
-import Control.Exception (try)
-import Control.Monad.State.Lazy (when)
 import Data.Aeson
   ( FromJSON,
     eitherDecode,
   )
 import qualified Data.ByteString.Lazy as L
-import qualified Data.ByteString.Lazy.Char8 as LC
 import qualified Data.Text as T
-import Data.Time (getCurrentTime)
 import GHC.Generics (Generic)
-import Network.HTTP.Simple
-  ( HttpException,
-    Request,
-    Response,
-    httpLBS,
+import Logger.Data
+  ( Priority,
+    logFile,
+    time,
   )
 import System.Exit (die)
-
--- Data type for the logger.
-data Priority = DEBUG | INFO | WARNING | ERROR
-  deriving (Show, Eq, Ord, Generic, FromJSON)
 
 -- Data type for the configuration file.
 data Configuration = Configuration
@@ -43,14 +33,6 @@ data Configuration = Configuration
     logOutput :: T.Text
   }
   deriving (Show, Generic, FromJSON)
-
--- Getting current time for the logger.
-time :: IO String
-time = take 19 . show <$> getCurrentTime
-
--- Name of the logfile.
-logFile :: String
-logFile = "log.log"
 
 -- Getting information from configuration file.
 getConfiguration :: IO Configuration
@@ -89,48 +71,3 @@ myToken = do
 
 messengerHost :: IO String
 messengerHost = (++ "/bot") <$> myHost
-
--- Logging level.
-logLevel :: IO Priority
-logLevel = priorityLevel <$> getConfiguration
-
--- Function writes information to log.
-writingLine :: Priority -> String -> IO ()
-writingLine lvl str = do
-  logLevel' <- logLevel
-  if lvl >= logLevel'
-    then do
-      t <- time
-      let string = t ++ " UTC   " ++ showLevel lvl ++ " - " ++ str
-      out <- logOutput <$> getConfiguration
-      case out of
-        "file" -> appendFile logFile $ string ++ "\n"
-        _ -> print string
-    else pure ()
-  where
-    showLevel val = case val of
-      DEBUG -> "DEBUG  "
-      INFO -> "INFO   "
-      WARNING -> "WARNING"
-      ERROR -> "ERROR  "
-
--- Function for connecting to the server.
-connectToServer :: Request -> Int -> IO (Response LC.ByteString)
-connectToServer req num = do
-  x <- try $ httpLBS req
-  writingLine DEBUG $ show req
-  case x of
-    Left e -> do
-      writingLine ERROR $ show (e :: HttpException)
-      when (num == 0) $ do
-        getCurrentTime >>= print
-        putStrLn "Connection Failure"
-        putStrLn "Trying to set a connection... "
-      threadDelay 1000000
-      connectToServer req (num + 1)
-    Right v -> do
-      writingLine DEBUG $ show v
-      when (num /= 0) $ do
-        getCurrentTime >>= print
-        putStrLn "Connection restored"
-      pure v
