+++ src/Lib.hs
@@ -3,25 +3,14 @@
 module Lib where
 
 import Config
-  ( Configuration
-      ( apiVKVersion,
-        defaultRepeats,
-        groupIdVK,
-        helpMess,
-        repeatMess
-      ),
-    Priority (..),
-    getConfiguration,
+  ( Configuration (helpMess),
     currentMessenger,
-    messengerHost,
-    myHost,
-    myToken,
-    writingLine,
+    getConfiguration,
   )
 import Control.Monad.State.Lazy
   ( MonadState (get, put),
     MonadTrans (lift),
-    StateT (runStateT),
+    StateT,
     evalStateT,
     replicateM_,
   )
@@ -29,122 +18,39 @@ import Data.Aeson
   ( encode,
   )
 import qualified Data.ByteString.Lazy.Char8 as LC
-import Data.Functor.Identity (runIdentity)
 import qualified Data.Map.Lazy as Map
 import qualified Data.Text as T
-import Data.Time (getCurrentTime)
 import Environment
   ( Environment (..),
     NumRepeats (..),
     UpdateID (..),
     Username (Username),
   )
-import KeyboardData
-  ( createKeyboard,
-    keyboardVk,
-  )
+import Logger.Data (Priority (..))
+import Logger.Functions (writingLine)
 import Network.HTTP.Simple
-  ( Request,
-    Response,
+  ( Response,
     httpLBS,
-    parseRequest_,
   )
-import System.Random
-  ( Random (randomRIO),
+import RequestBuilding
+  ( createQuestion,
+    createStringRequest,
+    getNumRepeats,
+    stringToUrl,
   )
-import Telegram
+import System.Random (Random (randomRIO))
+import Telegram.Data
   ( Chat (chat_id, username),
-    Media (AudioMessage, Others, Sticker, type_media),
-    Message (attachments, chat, message_id, textM),
+    Message (chat, message_id, textM),
     MessageDate (..),
     WholeObject (result),
     errorMessage,
-    getData,
   )
+import Telegram.Functions (getData)
+import Telegram.KeyboardData (createKeyboard)
 import Text.Read (readMaybe)
-
--- Request generation.
-createStringRequest :: String -> IO Request
-createStringRequest str = do
-  conf <- getConfiguration
-  crntMsngr <- currentMessenger
-  myHost' <- myHost
-  messengerHost' <- messengerHost
-  myToken' <- myToken
-  pure . parseRequest_ $
-    case crntMsngr of
-      "TG" -> mconcat ["https://", messengerHost', myToken', str]
-      _ ->
-        mconcat
-          [ "https://",
-            myHost',
-            "/method/messages.send?user_id=",
-            str,
-            "&peer_id=-",
-            show $ groupIdVK conf,
-            "&access_token=",
-            myToken',
-            "&v=",
-            T.unpack $ apiVKVersion conf
-          ]
-
--- Generating random numbers for VK requests.
-randomId :: IO Int
-randomId = randomRIO (1, 1000000)
-
--- Sending a request to VK to return a message.
-repeatMessageVk :: MessageDate -> IO (Response LC.ByteString)
-repeatMessageVk obj = do
-  r <- randomId
-  let userId = T.unpack $ username $ chat $ message obj
-      str = case textM $ message obj of
-        Just s -> s
-        _ -> ""
-      arr = case attachments $ message obj of
-        Just ls -> ls
-        _ -> []
-      add = case arr of
-        [] -> ""
-        _ -> case type_media $ (\(x : _) -> x) arr of
-          "sticker" -> ""
-          _ -> "&attachment="
-  string <-
-    createStringRequest $
-      mconcat
-        [ userId,
-          "&random_id=",
-          show r,
-          "&message=",
-          stringToUrl $ T.unpack str ++ add ++ attachment arr userId
-        ]
-  writingLine DEBUG $ show string
-  httpLBS string
-
--- Processing of attachments for VK.
-attachment :: [Media] -> String -> String
-attachment [] _ = ""
-attachment (x : xs) userId = case x of
-  Sticker _ n -> "&sticker_id=" ++ show n ++ attachment xs userId
-  AudioMessage _ l -> T.unpack l ++ attachment xs userId
-  Others t mI oI u k -> runIdentity $ do
-    let lnk = case u of
-          Just txt -> txt
-          _ -> ""
-    pure $
-      if t == "doc" && userId == show oI
-        then T.unpack lnk ++ "," ++ attachment xs userId
-        else
-          mconcat
-            [ T.unpack t,
-              show oI,
-              "_",
-              show mI,
-              case k of
-                Just s -> "_" ++ T.unpack s
-                _ -> "",
-              ",",
-              attachment xs userId
-            ]
+import Vk.Functions (repeatMessageVk)
+import Vk.KeyboardData (keyboardVk)
 
 -- Sending repetitions of request.
 sendRepeats :: MessageDate -> Environment -> IO ()
@@ -315,34 +221,9 @@ wordIsRepeat WorkHandle {..} getDataVk obj (x : xs) = do
           wordIsRepeatH WorkHandle {..} getDataVk obj xs
       )
 
-createQuestion :: MessageDate -> Environment -> IO String
-createQuestion obj env = do
-  conf <- getConfiguration
-  num <- getNumRepeats obj env
-  pure $
-    mconcat
-      [ "Currently set to ",
-        show num,
-        " repetitions.\n",
-        T.unpack $ repeatMess conf
-      ]
-
-stringToUrl :: String -> String
-stringToUrl = Prelude.foldl encodingChar ""
-  where
-    encodingChar acc c =
-      acc ++ case c of
-        '{' -> "%7B"
-        '}' -> "%7D"
-        '[' -> "%5B"
-        ']' -> "%5D"
-        '\"' -> "%22"
-        '\n' -> "%0A"
-        ' ' -> "%20"
-        ',' -> "%2C"
-        '.' -> "%2E"
-        ':' -> "%3A"
-        _ -> [c]
+-- Generating random numbers for VK requests.
+randomId :: IO Int
+randomId = randomRIO (1, 1000000)
 
 sendKeyboard :: MessageDate -> Environment -> IO (Response LC.ByteString)
 sendKeyboard obj env = do
@@ -396,51 +277,3 @@ sendComment obj str = do
           ]
   writingLine DEBUG $ show string
   httpLBS string
-
-getNumRepeats :: MessageDate -> Environment -> IO NumRepeats
-getNumRepeats obj env = do
-  conf <- getConfiguration
-  pure $ case Map.lookup usrName $ userData env of
-    Nothing -> NumRepeats $ defaultRepeats conf
-    Just n -> n
-  where
-    usrName = Username . username $ chat $ message obj
-
--- Function for getting update_id  for the first time.
---  (Excludes processing of messages sent before the program is started).
-firstUpdateIDSession :: StateT Environment IO ()
-firstUpdateIDSession = do
-  env <- get
-  (obj, newEnv) <- lift $ runStateT getData env
-  case obj of
-    Nothing -> pure ()
-    _ -> do
-      lift $ getCurrentTime >>= print
-      lift $ putStrLn "Connection established"
-      let update_id' = case result <$> obj of
-            Just [] -> 0
-            Just md -> (\(x : _) -> update_id x) (reverse md)
-            _ -> 0
-      if lastUpdate newEnv == 1
-        then put $ Environment update_id' (userData newEnv)
-        else put $ Environment (1 + update_id') (userData newEnv)
-
--- Main program cycle for Telegram.
-endlessCycle :: StateT Environment IO ()
-endlessCycle = do
-  env <- get
-  obj <- lift $ evalStateT getData env
-  let nothing _ = pure Nothing
-  case obj of
-    Nothing -> do
-      lift $ writingLine ERROR "Broken request!"
-      endlessCycle
-    _ -> do
-      let arr = case result <$> obj of
-            Just [x] -> [x]
-            _ -> []
-          update_id' = if null arr then 0 else (\(x : _) -> update_id x) (reverse arr)
-      put $ Environment (1 + update_id') (userData env)
-      _ <- get
-      mapM_ (ifKeyWord handler nothing) arr
-      endlessCycle
