+++ src/Vk/KeyboardData.hs
@@ -1,7 +1,7 @@
 {-# LANGUAGE DeriveAnyClass #-}
 {-# LANGUAGE DeriveGeneric #-}
 
-module KeyboardData where
+module Vk.KeyboardData where
 
 import Data.Aeson
   ( ToJSON (toJSON),
@@ -10,30 +10,6 @@ import Data.Aeson
   )
 import qualified Data.Text as T
 import GHC.Generics (Generic)
-
--- Data type for the Telegram keyboard.
-newtype KeyboardButton = KeyboardButton
-  {text :: T.Text}
-  deriving (Show, Generic, ToJSON)
-
-data ReplyKeyboardMarkup = ReplyKeyboardMarkup
-  { keyboard :: [[KeyboardButton]],
-    resize_keyboard :: Bool,
-    one_time_keyboard :: Bool
-  }
-  deriving (Show, Generic, ToJSON)
-
--- Creating keyboard for TG.
-toKeyboardButton :: Int -> KeyboardButton
-toKeyboardButton num = KeyboardButton {text = T.pack $ show num}
-
-createKeyboard :: ReplyKeyboardMarkup
-createKeyboard =
-  ReplyKeyboardMarkup
-    { keyboard = [map toKeyboardButton [1 .. 5]],
-      resize_keyboard = True,
-      one_time_keyboard = True
-    }
 
 -- Data types for the VK keyboard.
 data ActionVk = ActionVk
